+++ packages/dev/core/src/Physics/index.ts
@@ -3,3 +3,4 @@ export * from "./v1/index";
 export * from "./v2/index";
 export * from "./physicsEngineComponent";
 export * from "./v1/physicsEngineComponent";
+export * from "./physicsHelper";