+++ packages/dev/core/src/Physics/v2/IPhysicsEnginePlugin.ts
@@ -92,7 +92,7 @@ export interface MassProperties {
     /**
      *
      */
-    intertia: Vector3;
+    inertia: Vector3;
     /**
      *
      */
@@ -138,6 +138,7 @@ export interface IPhysicsEnginePluginV2 {
     setLinearVelocity(body: PhysicsBody, linVel: Vector3): void;
     getLinearVelocityToRef(body: PhysicsBody, linVel: Vector3): void;
     applyImpulse(body: PhysicsBody, location: Vector3, impulse: Vector3): void;
+    applyForce(body: PhysicsBody, location: Vector3, force: Vector3): void;
     setAngularVelocity(body: PhysicsBody, angVel: Vector3): void;
     getAngularVelocityToRef(body: PhysicsBody, angVel: Vector3): void;
     getBodyGeometry(body: PhysicsBody): {};